build(deps): bump project dependencies, follow up fixes after refactoring
- [x] update project dependencies to the latest versions;
- [x] rename unit-testing module, add types to the tools jest config;
