+++ functions/package.json
@@ -41,7 +41,7 @@
     "dotenv": "16.0.3",
     "express": "4.18.2",
     "firebase-admin": "11.1.0",
-    "firebase-functions": "3.24.1",
+    "firebase-functions": "4.0.0",
     "google-protobuf": "3.21.2",
     "graphql": "16.6.0",
     "graphql-subscriptions": "2.0.0",
