+++ libs/client-testing-e2e/src/index.ts
@@ -1,3 +1,2 @@
-// export * from './lib/client-testing-e2e';
 export * from './lib/support';
 export * from './lib/support/config/commands.config';
