+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nx-ng-starter",
-  "version": "5.5.0",
+  "version": "5.5.1",
   "private": true,
   "description": "Monorepo starter: Nx, Angular, Angular Elements, Electron, NodeJS, NestJS, Firebase.",
   "license": "MIT",
@@ -200,7 +200,7 @@
     "isomorphic-fetch": "3.0.0",
     "reflect-metadata": "0.1.13",
     "rxjs": "7.5.7",
-    "tslib": "^2.3.0",
+    "tslib": "2.4.0",
     "zone.js": "0.11.8"
   },
   "devDependencies": {
@@ -263,10 +263,10 @@
     "@types/google-protobuf": "3.15.6",
     "@types/graphql-upload": "8.0.11",
     "@types/hammerjs": "2.0.41",
-    "@types/jest": "28.1.1",
+    "@types/jest": "29.1.2",
     "@types/jsdom": "20.0.0",
     "@types/marked": "4.0.7",
-    "@types/node": "16.11.7",
+    "@types/node": "18.11.0",
     "@types/readline-sync": "1.4.4",
     "@types/websocket": "1.0.5",
     "@types/ws": "8.5.3",
@@ -296,20 +296,20 @@
     "eslint-plugin-unicorn": "44.0.2",
     "firebase": "9.12.1",
     "firebase-admin": "11.1.0",
-    "firebase-functions": "3.24.1",
+    "firebase-functions": "4.0.0",
     "google-protobuf": "3.21.2",
     "graphql-tag": "2.12.6",
     "hammerjs": "2.0.8",
     "husky": "8.0.1",
     "jasmine-marbles": "0.9.2",
-    "jest": "28.1.1",
-    "jest-environment-jsdom": "28.1.1",
+    "jest": "29.2.0",
+    "jest-environment-jsdom": "29.2.0",
     "jest-preset-angular": "12.2.2",
     "marked": "4.1.1",
     "material-design-icon-fonts-self-hosted-web": "1.0.7",
     "memo-decorator": "2.0.1",
     "mini-css-extract-plugin": "2.6.1",
-    "mocha": "10.0.0",
+    "mocha": "10.1.0",
     "mochawesome": "7.1.3",
     "mochawesome-merge": "4.2.1",
     "moment": "2.29.4",
@@ -333,10 +333,10 @@
     "stylelint-config-rational-order": "0.1.2",
     "stylelint-prettier": "2.0.0",
     "terser-webpack-plugin": "5.3.6",
-    "ts-jest": "28.0.5",
+    "ts-jest": "29.0.3",
     "ts-node": "10.9.1",
     "typescript": "4.8.4",
-    "util": "0.12.4",
+    "util": "0.12.5",
     "web-animations-js": "2.3.2",
     "webpack": "5.74.0",
     "webpack-bundle-analyzer": "4.6.1",
