+++ tools/jest.config.ts
@@ -1,4 +1,6 @@
-const config = {
+import { Config } from '@jest/types';
+
+const config: Config.InitialOptions = {
   coverageDirectory: '../coverage/tools',
   coverageThreshold: {
     global: {
